+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.2.2-dev:
 
 - refactor main API/update logic from NavBar into Brain
+- consolidate client/server types
 
 v0.2.1: